2019.6.17  11.05

+++ src/com/spring/web/model/DangerCoordinate.java
@@ -4,25 +4,25 @@ public class DangerCoordinate {
 
     private Integer id;
     private Integer user_id;
-    private Integer danger1;
-    private Integer danger2;
-    private Integer danger3;
-    private Integer danger4;
-    private Integer danger5;
-    private Integer danger6;
-    private Integer danger7;
-    private Integer danger8;
-    private Integer danger9;
-    private Integer danger10;
-    private Integer danger11;
-    private Integer danger12;
-    private Integer danger13;
-    private Integer counts;
+    private Double danger1;
+    private Double danger2;
+    private Double danger3;
+    private Double danger4;
+    private Double danger5;
+    private Double danger6;
+    private Double danger7;
+    private Double danger8;
+    private Double danger9;
+    private Double danger10;
+    private Double danger11;
+    private Double danger12;
+    private Double danger13;
+    private Double counts;
 
     public DangerCoordinate() {
     }
 
-    public DangerCoordinate(Integer id, Integer user_id, Integer danger1, Integer danger2, Integer danger3, Integer danger4, Integer danger5, Integer danger6, Integer danger7, Integer danger8, Integer danger9, Integer danger10, Integer danger11, Integer danger12, Integer danger13, Integer counts) {
+    public DangerCoordinate(Integer id, Integer user_id, Double danger1, Double danger2, Double danger3, Double danger4, Double danger5, Double danger6, Double danger7, Double danger8, Double danger9, Double danger10, Double danger11, Double danger12, Double danger13, Double counts) {
         this.id = id;
         this.user_id = user_id;
         this.danger1 = danger1;
@@ -57,115 +57,115 @@ public class DangerCoordinate {
         this.user_id = user_id;
     }
 
-    public Integer getDanger1() {
+    public Double getDanger1() {
         return danger1;
     }
 
-    public void setDanger1(Integer danger1) {
+    public void setDanger1(Double danger1) {
         this.danger1 = danger1;
     }
 
-    public Integer getDanger2() {
+    public Double getDanger2() {
         return danger2;
     }
 
-    public void setDanger2(Integer danger2) {
+    public void setDanger2(Double danger2) {
         this.danger2 = danger2;
     }
 
-    public Integer getDanger3() {
+    public Double getDanger3() {
         return danger3;
     }
 
-    public void setDanger3(Integer danger3) {
+    public void setDanger3(Double danger3) {
         this.danger3 = danger3;
     }
 
-    public Integer getDanger4() {
+    public Double getDanger4() {
         return danger4;
     }
 
-    public void setDanger4(Integer danger4) {
+    public void setDanger4(Double danger4) {
         this.danger4 = danger4;
     }
 
-    public Integer getDanger5() {
+    public Double getDanger5() {
         return danger5;
     }
 
-    public void setDanger5(Integer danger5) {
+    public void setDanger5(Double danger5) {
         this.danger5 = danger5;
     }
 
-    public Integer getDanger6() {
+    public Double getDanger6() {
         return danger6;
     }
 
-    public void setDanger6(Integer danger6) {
+    public void setDanger6(Double danger6) {
         this.danger6 = danger6;
     }
 
-    public Integer getDanger7() {
+    public Double getDanger7() {
         return danger7;
     }
 
-    public void setDanger7(Integer danger7) {
+    public void setDanger7(Double danger7) {
         this.danger7 = danger7;
     }
 
-    public Integer getDanger8() {
+    public Double getDanger8() {
         return danger8;
     }
 
-    public void setDanger8(Integer danger8) {
+    public void setDanger8(Double danger8) {
         this.danger8 = danger8;
     }
 
-    public Integer getDanger9() {
+    public Double getDanger9() {
         return danger9;
     }
 
-    public void setDanger9(Integer danger9) {
+    public void setDanger9(Double danger9) {
         this.danger9 = danger9;
     }
 
-    public Integer getDanger10() {
+    public Double getDanger10() {
         return danger10;
     }
 
-    public void setDanger10(Integer danger10) {
+    public void setDanger10(Double danger10) {
         this.danger10 = danger10;
     }
 
-    public Integer getDanger11() {
+    public Double getDanger11() {
         return danger11;
     }
 
-    public void setDanger11(Integer danger11) {
+    public void setDanger11(Double danger11) {
         this.danger11 = danger11;
     }
 
-    public Integer getDanger12() {
+    public Double getDanger12() {
         return danger12;
     }
 
-    public void setDanger12(Integer danger12) {
+    public void setDanger12(Double danger12) {
         this.danger12 = danger12;
     }
 
-    public Integer getDanger13() {
+    public Double getDanger13() {
         return danger13;
     }
 
-    public void setDanger13(Integer danger13) {
+    public void setDanger13(Double danger13) {
         this.danger13 = danger13;
     }
 
-    public Integer getCounts() {
+    public Double getCounts() {
         return counts;
     }
 
-    public void setCounts(Integer counts) {
+    public void setCounts(Double counts) {
         this.counts = counts;
     }
 }
